Fix character handling in enter.c
The inputted CTRL('c') characters are handled on _this_ round of the
main while loop, not on the next.

Inputted characters are now printed to screen (echo(), noecho()).

Cursor is enabled before stating the input, and hidden after it.

Input line is cleared after use (i.e. before returning from the
function.